Support default values for variables
