+++ dev/com.ibm.ws.config/src/com/ibm/ws/config/schemagen/internal/SchemaWriter.java
@@ -1318,9 +1318,17 @@ class SchemaWriter {
         writer.writeStartElement(XSD, "attribute");
         writer.writeAttribute("name", "value");
         writer.writeAttribute("type", "xsd:string");
-        writer.writeAttribute("use", "required");
+
         doc = resourceBundle.getString("config.internal.metatype.variableDefinitionType.value.documentation");
         label = resourceBundle.getString("config.internal.metatype.variableDefinitionType.value.label");
+        writeDocumentation(doc, label);
+        writer.writeEndElement();
+
+        writer.writeStartElement(XSD, "attribute");
+        writer.writeAttribute("name", "defaultValue");
+        writer.writeAttribute("type", "xsd:string");
+        doc = resourceBundle.getString("config.internal.metatype.variableDefinitionType.defaultValue.documentation");
+        label = resourceBundle.getString("config.internal.metatype.variableDefinitionType.defaultValue.label");
         writeDocumentation(doc, label);
         writer.writeEndElement();
