+++ dev/com.ibm.ws.config_fat/publish/servers/com.ibm.ws.config.delayedVar/server.env
@@ -1,0 +1,4 @@
+fromEnvironment=envValue
+FROMENVIRONMENTCAPS=caps
+from__environment=mangled
+FROM_ENVIRONMENT_CAPS=mangledCaps
