+++ dev/com.ibm.ws.config_fat/publish/servers/com.ibm.ws.config.delayedVar/server.xml
@@ -19,6 +19,8 @@
 		type="war" location="varmerge.war" />
 
 	<variable name="variableDelayTest" value="evaluated" />
+	
+	<variable name="variableDefaultTest" defaultValue="this is the default"/>
 
 	<delayedVar delayedVariable="${variableDelayTest}"
 		immediateVar="${variableDelayTest}" immediateVarTwo="${variableDelayTest}" />
