+++ dev/com.ibm.ws.config_fat/test-bundles/test.config.variables/src/com/ibm/ws/config/variables/ServerXMLVariables.java
@@ -18,4 +18,6 @@ import java.util.Map;
 public interface ServerXMLVariables {
 
     public Map<String, String> getServerXMLVariables();
+
+    public Map<String, String> getServerXMLVariableDefaultValues();
 }
